requirements.txt
requests
python-dotenv
pandas
google-api-python-client
google-auth
google-auth-oauthlib
tweepy
praw